requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pytest>=7.0.0
pikepdf>=8.0.0
pdfplumber>=0.10.0
sentence-transformers>=2.0.0
chromadb>=0.4.0
